+++ src/main/java/org/folio/sidecar/service/filter/DesiredPermissionsFilter.java
@@ -74,7 +74,7 @@ public class DesiredPermissionsFilter implements IngressRequestFilter {
     }
 
     putPermissionsHeaderToRequest(rc, permissions);
-    log.info("X-Okapi-Permissions populated: permissions = {}, userId = {}", permissions, userId);
+    log.debug("X-Okapi-Permissions populated: permissions = {}, userId = {}", permissions, userId);
     return rc;
   }