+++ NEWS.md
@@ -18,3 +18,4 @@
 * Fix secure store env variable processing in native sidecar (MODSIDECAR-145)
 * Add key store configuration properties for secure store (APPPOCTOOL-62)
 * Add support for custom Keycloak base URL for JWKS endpoint, `KC_JWKS_BASE_URL` environment variable (MODSIDECAR-148)
+* Review and clean up logs in folio-module-sidecar (MODSIDECAR-164)